+++ android/history.md
@@ -1,6 +1,7 @@
 # Keyman for Android
 
 ## 12.0 alpha
+* Add feature to install lexical model packages (#1641)
 * Started work on Keyman for Android 12.
 
 ## 2019-02-27 11.0.2102 stable